[FLINK-3049] move Either type to org.apache.flink.types
